# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feedback:
    get:
      operationId: listFeedback
      tags: [feedback]
      summary: List public teacher feedback
      responses:
        "200":
          description: Recent teacher feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"
    post:
      operationId: createFeedback
      tags: [feedback]
      summary: Submit teacher feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Saved feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
        "400":
          description: Invalid feedback
        "409":
          description: A very similar feedback submission was already received recently
        "429":
          description: Too many feedback submissions from the same source
          headers:
            Retry-After:
              description: Seconds until another submission may be attempted
              schema:
                type: integer
...
  /feedback/manage:
    get:
      operationId: listManagedFeedback
      tags: [feedback]
      summary: List feedback for account management
      responses:
        "200":
          description: All teacher feedback with visibility state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedFeedback"
        "401":
          description: Authentication required
        "403":
          description: Insufficient role
  /feedback/{feedbackId}:
    patch:
      operationId: updateFeedback
      tags: [feedback]
      summary: Change feedback visibility
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackUpdate"
      responses:
        "200":
          description: Updated feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedFeedback"
        "401":
          description: Authentication required
        "403":
          description: Insufficient role
        "404":
          description: Feedback not found
    delete:
      operationId: deleteFeedback
      tags: [feedback]
      summary: Delete feedback
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
      responses:
        "204":
          description: Deleted
        "401":
          description: Authentication required
        "403":
          description: Admin role required
        "404":
          description: Feedback not found
  /feedback/{feedbackId}/history:
    get:
      operationId: listFeedbackHistory
      tags: [feedback]
      summary: List feedback moderation history
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
      responses:
        "200":
          description: Moderation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackModerationHistory"
        "401":
          description: Authentication required
        "403":
          description: Insufficient role
        "404":
          description: Feedback not found
  /feedback/{feedbackId}/report:
    post:
      operationId: createFeedbackReport
      tags: [feedback]
      summary: Report public feedback
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackReportInput"
      responses:
        "201":
          description: Feedback report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReport"
        "400":
          description: Invalid report
        "404":
          description: Feedback not found
        "409":
          description: A matching open report already exists
  /feedback/reports:
    get:
      operationId: listFeedbackReports
      tags: [feedback]
      summary: List feedback reports for moderators
      responses:
        "200":
          description: Feedback reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackReport"
        "401":
          description: Authentication required
        "403":
          description: Insufficient role
  /feedback/reports/{reportId}:
    patch:
      operationId: updateFeedbackReport
      tags: [feedback]
      summary: Resolve a feedback report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackReportUpdate"
      responses:
        "200":
          description: Updated feedback report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReport"
        "400":
          description: Invalid report status
        "401":
          description: Authentication required
        "403":
          description: Insufficient role
        "404":
          description: Report not found
...
  /session/me:
    get:
      operationId: getCurrentUser
      tags: [session]
      summary: Get the current application user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "401":
          description: Authentication required
...
components:
  parameters:
...
    FeedbackId:
      name: feedbackId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Feedback:
      type: object
      required: [id, teacherName, feedback, rating, createdAt]
      properties:
        id: { type: integer }
        teacherName: { type: string }
        feedback: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        createdAt: { type: string, format: date-time }
    FeedbackInput:
      type: object
      required: [teacherName, feedback, rating]
      properties:
        teacherName: { type: string, minLength: 2, maxLength: 80 }
        feedback: { type: string, minLength: 3, maxLength: 500 }
        rating: { type: integer, minimum: 1, maximum: 5 }
...
    ManagedFeedback:
      allOf:
        - $ref: "#/components/schemas/Feedback"
        - type: object
          required: [isVisible]
          properties:
            isVisible: { type: boolean }
    FeedbackUpdate:
      type: object
      required: [isVisible]
      properties:
        isVisible: { type: boolean }
        reason: { type: [string, "null"], maxLength: 500 }
    FeedbackModerationHistory:
      type: object
      required: [id, feedbackId, action, actorId, actorName, createdAt]
      properties:
        id: { type: integer }
        feedbackId: { type: integer }
        action:
          type: string
          enum: [published, hidden, deleted]
        reason: { type: [string, "null"] }
        actorId: { type: string }
        actorName: { type: string }
        createdAt: { type: string, format: date-time }
    FeedbackReportInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum: [spam, offensive, advertising, duplicate, other]
        details: { type: [string, "null"], maxLength: 500 }
    FeedbackReport:
      type: object
      required: [id, feedbackId, reason, status, createdAt]
      properties:
        id: { type: integer }
        feedbackId: { type: integer }
        reason:
          type: string
          enum: [spam, offensive, advertising, duplicate, other]
        details: { type: [string, "null"] }
        status:
          type: string
          enum: [open, reviewed, dismissed]
        reviewedBy: { type: [string, "null"] }
        reviewedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    FeedbackReportUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [reviewed, dismissed]
...
    AppUser:
      type: object
      required: [id, name, email, role, joinedAt, lastActiveAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        avatarUrl: { type: ["string", "null"], format: uri }
        role:
          type: string
          enum: [admin, moderator, member]
        joinedAt: { type: string, format: date-time }
        lastActiveAt: { type: string, format: date-time }
        profileVisibility:
          type: string
          enum: [private, public]
...